Auto retry to killChild
I am seeing a specific case where a job starts at the exact same time as it receives a shutdown signal.  The child goes on to start the job and never gets requeued when it gets hard killed a moment later.  My best guess is it is from trying to kill the child before it is fully started.  I do not have a reliable way of reproducing this issue outside of production